issue #21 show reminders tab
