+++ src/js/consts/index.js
@@ -1,3 +1,4 @@
 export * from "./engine";
 export * from "./message";
 export * from "./storage";
+export * from "./optionsTabs";
