+++ src/js/consts/optionsTabs.js
@@ -1,0 +1,7 @@
+const OPTIONS_TAB_HASH_SKILLS_IMPORTANCE = "#skills-importance";
+const OPTIONS_TAB_HASH_REMINDERS = "#reminders";
+
+export {
+    OPTIONS_TAB_HASH_SKILLS_IMPORTANCE,
+    OPTIONS_TAB_HASH_REMINDERS,
+};
