+++ src/js/page/Popup/Popup.js
@@ -1,50 +1,30 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import Button from "@/component/Button";
-import ReminderList from "@/component/ReminderList";
-import { getAllReminders } from "@/service/ReminderService";
+import { OPTIONS_TAB_HASH_SKILLS_IMPORTANCE, OPTIONS_TAB_HASH_REMINDERS } from "@/consts";
 import "./Popup.scss";
 
-const handleSkillsButtonClick = () => chrome.runtime.openOptionsPage();
+const handleOptionsClick = (section) => window.open(
+    `${chrome.runtime.getURL("options-page.html")}${section}`,
+);
 
 const handleSokkerButtonClick = () => chrome.tabs.create({ url: "https://sokker.org" });
 
-const PopupPage = () => {
-    const [reminders, setReminders] = useState([]);
+const PopupPage = () => (
+    <div className="popup-page">
+        <h1>Sokker Position Rating</h1>
 
-    useEffect(async () => setReminders(await getAllReminders()), []);
+        <div className="popup-page__controls">
+            <Button
+                onClick={() => handleOptionsClick(OPTIONS_TAB_HASH_SKILLS_IMPORTANCE)}
+            >
+                Set skills importance
+            </Button>
 
-    const handleReminderRemove = (reminder) => {
-        const index = reminders.findIndex(r => r.player === reminder.player && r.remindDate === reminder.remindDate);
+            <Button onClick={() => handleOptionsClick(OPTIONS_TAB_HASH_REMINDERS)}>Show active reminders</Button>
 
-        if (index > -1) {
-            const newReminders = reminders.slice();
-            newReminders.splice(index, 1);
-            setReminders(newReminders);
-        }
-    };
-
-    return (
-        <div className="popup-page">
-            <h1>Sokker Position Rating</h1>
-
-            <div className="popup-page__controls">
-                <Button onClick={handleSkillsButtonClick}>Set skills importance</Button>
-                <Button onClick={handleSokkerButtonClick}>Go to sokker.org</Button>
-            </div>
-
-            {reminders.length > 0 && (
-                <div className="popup-page__reminders">
-                    <h3 className="popup-page__reminders-header">Active reminders:</h3>
-                    <ReminderList
-                        reminders={reminders}
-                        showLabel={false}
-                        showPlayerNames
-                        onRemove={handleReminderRemove}
-                    />
-                </div>
-            )}
+            <Button onClick={handleSokkerButtonClick}>Go to sokker.org</Button>
         </div>
-    );
-};
+    </div>
+);
 
 export default PopupPage;
